+++ src/factiva/core/__init__.py
@@ -3,10 +3,10 @@
 """
 __all__ = ['const', 'dicts', 'auth', 'UserKey', 'APIKeyUser', 'StreamUser', 'StreamResponse']
 
-from factiva.core.__version__ import __version__
-from factiva.core.auth.userkey import (UserKey)
-from factiva.core.apikeyuser import (APIKeyUser)
-from factiva.core.streamuser import (StreamUser)
-from factiva.core.stream_response import (StreamResponse)
+from .__version__ import __version__
+from .auth.userkey import (UserKey)
+from .apikeyuser import (APIKeyUser)
+from .streamuser import (StreamUser)
+from .stream_response import (StreamResponse)
 
 version = __version__
