+++ src/factiva/core/apikeyuser.py
@@ -3,7 +3,8 @@
 import requests
 import pandas as pd
 
-from factiva.helper import load_environment_value, mask_string, api_send_request, flatten_dict
+from .req import api_send_request
+from .tools import load_environment_value, mask_string, flatten_dict
 from factiva.core import const
