+++ src/factiva/core/req/__init__.py
@@ -1,25 +1,10 @@
-"""Implement different helper functions."""
-import os
-import json
-import requests
-from factiva.core import const
-
-@DeprecationWarning('Will be removed in 1.0')
-def load_environment_value(config_key):
-    """Obtain a environmental variable."""
-    tmp_val = os.getenv(config_key, None)
-    if tmp_val is None:
-        raise Exception("Environment Variable {} not found!".format(config_key))
-    return tmp_val
+"""
+    Module to handle the API and other requests
+"""
+import requests, json
+from .. import const
 
 
-@DeprecationWarning('Will be removed in 1.0')
-def mask_string(raw_str, right_padding=4):
-    """Mask a string with a default value of 4."""
-    return raw_str[-right_padding:].rjust(len(raw_str), '*')
-
-
-@DeprecationWarning('Will be removed in 1.0')
 def send_get_request(endpoint_url=const.API_HOST, headers=None, qs_params=None):
     """Send get request."""
     if (qs_params is not None) and (not isinstance(qs_params, dict)):
@@ -27,7 +12,6 @@ def send_get_request(endpoint_url=const.API_HOST, headers=None, qs_params=None):
     return requests.get(endpoint_url, headers=headers, params=qs_params)
 
 
-@DeprecationWarning('Will be removed in 1.0')
 def send_post_request(endpoint_url=const.API_HOST, headers=None, payload=None):
     """Send post request."""
     if payload is not None:
@@ -42,7 +26,6 @@ def send_post_request(endpoint_url=const.API_HOST, headers=None, payload=None):
     return requests.post(endpoint_url, headers=headers)
 
 
-@DeprecationWarning('Will be removed in 1.0')
 def api_send_request(method='GET', endpoint_url=const.API_HOST, headers=None, payload=None, qs_params=None):
     """Send a generic request to a certain API end point."""
     if headers is None:
@@ -68,35 +51,3 @@ def api_send_request(method='GET', endpoint_url=const.API_HOST, headers=None, pa
         raise RuntimeError('API Request failed. Unspecified Error.')
 
     return response
-
-
-@DeprecationWarning('Will be removed in 1.0')
-def validate_type(var_to_validate, expected_type, error_message):
-    """Validate a given type."""
-    if not isinstance(var_to_validate, expected_type):
-        raise ValueError(error_message)
-
-
-@DeprecationWarning('Will be removed in 1.0')
-def flatten_dict(multi_level_dict):
-    """Flatten a dictionary."""
-    flattened_dict = {}
-    for entry in multi_level_dict:
-        if isinstance(multi_level_dict[entry], dict):
-            new_elements = flatten_dict(multi_level_dict[entry])
-            flattened_dict.update(new_elements)
-        else:
-            flattened_dict[entry] = multi_level_dict[entry]
-
-    return flattened_dict
-
-# Apparently unused
-# def load_generic_env_variable(variable, env_var):
-#     """Load generic env variable."""
-#     if not variable:
-#         try:
-#             return load_environment_value(f'{env_var}')
-#         except Exception:
-#             return variable
-
-#     return variable
